Implement basic float expr and printing
