+++ compiler/modules/front.syntax/astString.bal
@@ -241,6 +241,9 @@ function exprToWords(Word[] w, Expr expr, boolean wrap = false) {
             w.push(expr.digits);
         }
     }
+    else if expr is FpLiteralExpr {
+        w.push(expr.untypedLiteral, CLING, expr.typeSuffix);
+    }
     else if expr is UnaryExpr {
         if wrap {
             w.push("(");
