+++ compiler/modules/front.syntax/tests/data/V-stmt-foreach-trailing-dot.bal
@@ -1,0 +1,11 @@
+function wrapper(boolean b) {
+    // @case
+    foreach int i in 0..<1 {
+        foo(i);
+    }
+    // @end
+}
+
+
+function foo(int a) {
+}
