+++ compiler/testSuite/Vfloat00.bal
@@ -1,0 +1,5 @@
+import ballerina/io;
+public function main() {
+    float f = 0.1;
+    io:println(f); // @output 0.1
+}
