+++ compiler/tests/testTypes.bal
@@ -1,0 +1,7 @@
+import ballerina/test;
+
+// XXX move testJsonTypes here
+@test:Config {}
+function testTypesJson() returns error? {
+    check testJsonTypes("tests.json");
+}
